requirements.txt
torch>=1.12.0
transformers>=4.20.0
numpy>=1.21.0
scikit-learn>=1.1.0
datasets>=2.0.0
accelerate>=0.12.0
hydra-core>=1.3.0
omegaconf>=2.3.0
rich>=13.0.0
matplotlib>=3.5.0
seaborn>=0.11.0